Unified unions tests
