+++ modules/doobie/src/test/scala/unions/UnionsMapping.scala
@@ -1,0 +1,22 @@
+// Copyright (c) 2016-2020 Association of Universities for Research in Astronomy, Inc. (AURA)
+// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+
+package unions
+
+import cats.effect.Sync
+import doobie.util.transactor.Transactor
+
+import edu.gemini.grackle.Mapping
+import edu.gemini.grackle.doobie.{DoobieMappingCompanion, DoobieMonitor}
+
+import utils.{DatabaseSuite, DoobieTestMapping}
+import grackle.test.SqlUnionSpec
+
+object UnionsMapping extends DoobieMappingCompanion {
+  def mkMapping[F[_]: Sync](transactor: Transactor[F], monitor: DoobieMonitor[F]): Mapping[F] =
+    new DoobieTestMapping[F](transactor, monitor) with SqlUnionsMapping[F]
+}
+
+final class UnionsSpec extends DatabaseSuite with SqlUnionSpec {
+  lazy val mapping = UnionsMapping.mkMapping(xa)
+}
